Add channels for number of remaining flags.
The experience with the 9 channels used so far is that the model strongly prefers normal moves over flag moves. Very probably that is b/c during training it sees far more examples that make normal moves. And the current board representation makes it hard for the model to learn whether flags are available.